refactor(InputValidator): InputValidator삭제 후 검증기능 InputView로 전가
